Settings are now saved.
So as long as we save settings in "user://" they will persist on the web.

Welp, here we go. I created a ConfigFileHandler singleton to manage this, I then created a Settings_Loader which will load and set all appropriate settings on MainMenu load, I then plumbed up SettingsMenu to actually adjust the save file properly.

In theory this works.