HUD Now Complete
Made the font and coloring of the font conform to palette.

Made the hand health counter bloody...

Removed the sprite sheet in favor of individual images. This allows me to use Animation Player with just a basic sprite, so its *really* easy.

Added 6 Animations with Animation Player, changing the sprite and turning on and off the individual particle emitters to simulate blood. Well I had no sprite particle to make it better so it is what it is.

I also added a delay if we lose so you can see for a lack of a better term "Stump". I set that timer to 2 seconds, can adjust. May need logic to prevent further play in the game, maybe a whimper sound effect could play... WWise is not my department, so I'm not adding that.