Credits Screen Implemented + Font Change, UI Cleanup
Today I learned Margin Containers exist. They can be used as sperators.

Also I found a non-cringe CC0 Font that is easier on the eys, so I used it.

Finally I implemented the Credits Scene + Mandatory link to Godot's License page, which satisfies the requirements that we link to it.